hls client: merge fmp4 init file reader and writer
