+++ internal/hls/fmp4/init_test.go
@@ -46,9 +46,22 @@ var testAudioTrack = &gortsplib.TrackMPEG4Audio{
 	IndexDeltaLength: 3,
 }
 
-func TestInitWrite(t *testing.T) {
+func TestInitMarshal(t *testing.T) {
 	t.Run("video + audio", func(t *testing.T) {
-		byts, err := InitWrite(testVideoTrack, testAudioTrack)
+		init := Init{
+			VideoTrack: &InitTrack{
+				ID:        1,
+				TimeScale: 90000,
+				Track:     testVideoTrack,
+			},
+			AudioTrack: &InitTrack{
+				ID:        2,
+				TimeScale: uint32(testAudioTrack.ClockRate()),
+				Track:     testAudioTrack,
+			},
+		}
+
+		byts, err := init.Marshal()
 		require.NoError(t, err)
 
 		boxes := []gomp4.BoxPath{
@@ -173,7 +186,15 @@ func TestInitWrite(t *testing.T) {
 	})
 
 	t.Run("video only", func(t *testing.T) {
-		byts, err := InitWrite(testVideoTrack, nil)
+		init := Init{
+			VideoTrack: &InitTrack{
+				ID:        1,
+				TimeScale: 90000,
+				Track:     testVideoTrack,
+			},
+		}
+
+		byts, err := init.Marshal()
 		require.NoError(t, err)
 
 		boxes := []gomp4.BoxPath{
@@ -245,7 +266,15 @@ func TestInitWrite(t *testing.T) {
 	})
 
 	t.Run("audio only", func(t *testing.T) {
-		byts, err := InitWrite(nil, testAudioTrack)
+		init := &Init{
+			AudioTrack: &InitTrack{
+				ID:        1,
+				TimeScale: uint32(testAudioTrack.ClockRate()),
+				Track:     testAudioTrack,
+			},
+		}
+
+		byts, err := init.Marshal()
 		require.NoError(t, err)
 
 		boxes := []gomp4.BoxPath{
@@ -315,4 +344,118 @@ func TestInitWrite(t *testing.T) {
 		}
 		testMP4(t, byts, boxes)
 	})
+}
+
+func TestInitUnmarshal(t *testing.T) {
+	byts := []byte{
+		0x00, 0x00, 0x00, 0x1c, 0x66, 0x74, 0x79, 0x70,
+		0x64, 0x61, 0x73, 0x68, 0x00, 0x00, 0x00, 0x01,
+		0x69, 0x73, 0x6f, 0x6d, 0x61, 0x76, 0x63, 0x31,
+		0x64, 0x61, 0x73, 0x68, 0x00, 0x00, 0x02, 0x92,
+		0x6d, 0x6f, 0x6f, 0x76, 0x00, 0x00, 0x00, 0x6c,
+		0x6d, 0x76, 0x68, 0x64, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x98, 0x96, 0x80, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x01, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00,
+		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+		0x40, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x00, 0x00, 0x00, 0xff, 0xff, 0xff, 0xff,
+		0x00, 0x00, 0x01, 0xf6, 0x74, 0x72, 0x61, 0x6b,
+		0x00, 0x00, 0x00, 0x5c, 0x74, 0x6b, 0x68, 0x64,
+		0x00, 0x00, 0x00, 0x07, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00,
+		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+		0x40, 0x00, 0x00, 0x00, 0x03, 0xc0, 0x00, 0x00,
+		0x02, 0x1c, 0x00, 0x00, 0x00, 0x00, 0x01, 0x92,
+		0x6d, 0x64, 0x69, 0x61, 0x00, 0x00, 0x00, 0x20,
+		0x6d, 0x64, 0x68, 0x64, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x98, 0x96, 0x80, 0x00, 0x00, 0x00, 0x00,
+		0x55, 0xc4, 0x00, 0x00, 0x00, 0x00, 0x00, 0x38,
+		0x68, 0x64, 0x6c, 0x72, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x00, 0x00, 0x00, 0x76, 0x69, 0x64, 0x65,
+		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x00, 0x00, 0x00, 0x42, 0x72, 0x6f, 0x61,
+		0x64, 0x70, 0x65, 0x61, 0x6b, 0x20, 0x56, 0x69,
+		0x64, 0x65, 0x6f, 0x20, 0x48, 0x61, 0x6e, 0x64,
+		0x6c, 0x65, 0x72, 0x00, 0x00, 0x00, 0x01, 0x32,
+		0x6d, 0x69, 0x6e, 0x66, 0x00, 0x00, 0x00, 0x14,
+		0x76, 0x6d, 0x68, 0x64, 0x00, 0x00, 0x00, 0x01,
+		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x00, 0x00, 0x24, 0x64, 0x69, 0x6e, 0x66,
+		0x00, 0x00, 0x00, 0x1c, 0x64, 0x72, 0x65, 0x66,
+		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01,
+		0x00, 0x00, 0x00, 0x0c, 0x75, 0x72, 0x6c, 0x20,
+		0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0xf2,
+		0x73, 0x74, 0x62, 0x6c, 0x00, 0x00, 0x00, 0xa6,
+		0x73, 0x74, 0x73, 0x64, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x96,
+		0x61, 0x76, 0x63, 0x31, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x00, 0x00, 0x00, 0x03, 0xc0, 0x02, 0x1c,
+		0x00, 0x48, 0x00, 0x00, 0x00, 0x48, 0x00, 0x00,
+		0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x04, 0x68,
+		0x32, 0x36, 0x34, 0x00, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x18,
+		0xff, 0xff, 0x00, 0x00, 0x00, 0x30, 0x61, 0x76,
+		0x63, 0x43, 0x01, 0x42, 0xc0, 0x1f, 0xff, 0xe1,
+		0x00, 0x19, 0x67, 0x42, 0xc0, 0x1f, 0xd9, 0x00,
+		0xf0, 0x11, 0x7e, 0xf0, 0x11, 0x00, 0x00, 0x03,
+		0x00, 0x01, 0x00, 0x00, 0x03, 0x00, 0x30, 0x8f,
+		0x18, 0x32, 0x48, 0x01, 0x00, 0x04, 0x68, 0xcb,
+		0x8c, 0xb2, 0x00, 0x00, 0x00, 0x10, 0x70, 0x61,
+		0x73, 0x70, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00,
+		0x00, 0x01, 0x00, 0x00, 0x00, 0x10, 0x73, 0x74,
+		0x74, 0x73, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x00, 0x00, 0x00, 0x00, 0x10, 0x73, 0x74,
+		0x73, 0x63, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x00, 0x00, 0x00, 0x00, 0x14, 0x73, 0x74,
+		0x73, 0x7a, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x10, 0x73, 0x74, 0x63, 0x6f, 0x00, 0x00,
+		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x28, 0x6d, 0x76, 0x65, 0x78, 0x00, 0x00,
+		0x00, 0x20, 0x74, 0x72, 0x65, 0x78, 0x00, 0x00,
+		0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00,
+		0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+		0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+	}
+
+	var init Init
+	err := init.Unmarshal(byts)
+	require.NoError(t, err)
+
+	require.Equal(t, Init{
+		VideoTrack: &InitTrack{
+			ID:        256,
+			TimeScale: 10000000,
+			Track: &gortsplib.TrackH264{
+				PayloadType: 96,
+				SPS: []byte{
+					0x67, 0x42, 0xc0, 0x1f, 0xd9, 0x00, 0xf0, 0x11,
+					0x7e, 0xf0, 0x11, 0x00, 0x00, 0x03, 0x00, 0x01,
+					0x00, 0x00, 0x03, 0x00, 0x30, 0x8f, 0x18, 0x32,
+					0x48,
+				},
+				PPS: []byte{
+					0x68, 0xcb, 0x8c, 0xb2,
+				},
+			},
+		},
+	}, init)
 }
